Разбивает редьюсер, заменяет прямое получение данных из стейта в mapStateToProps на селекторы
