+++ src/reducer/data/data.js
@@ -1,0 +1,51 @@
+import {parseMovie, parseMovies} from "../../adapters/movies";
+import {extendObject} from "../../const";
+
+const initialState = {
+  movies: [],
+  promo: undefined,
+};
+
+const ActionType = {
+  LOAD_MOVIES: `LOAD_MOVIES`,
+  LOAD_PROMO: `LOAD_PROMO`,
+};
+
+const ActionCreator = {
+  loadMovies: (movies) => ({
+    type: ActionType.LOAD_MOVIES,
+    payload: movies,
+  }),
+  loadPromo: (promo) => ({
+    type: ActionType.LOAD_PROMO,
+    payload: promo,
+  }),
+};
+
+const Operation = {
+  loadMovies: () => (dispatch, getState, api) => {
+    return api.get(`/films`)
+      .then((response) => {
+        dispatch(ActionCreator.loadMovies(parseMovies(response.data)));
+      });
+  },
+  loadPromo: () => (dispatch, getState, api) => {
+    return api.get(`/films/promo`)
+      .then((response) => {
+        dispatch(ActionCreator.loadPromo(parseMovie(response.data)));
+      });
+  }
+};
+
+const reducer = (state = initialState, action) => {
+  switch (action.type) {
+    case ActionType.LOAD_MOVIES:
+      return extendObject(state, {movies: action.payload});
+    case ActionType.LOAD_PROMO:
+      return extendObject(state, {promo: action.payload});
+  }
+
+  return state;
+};
+
+export {reducer, ActionType, ActionCreator, Operation};
