+++ src/reducer/data/selectors.js
@@ -1,0 +1,11 @@
+import NameSpace from "../name-space";
+
+const NAME_SPACE = NameSpace.DATA;
+
+export const getAllMovies = (state) => {
+  return state[NAME_SPACE].movies;
+};
+
+export const getPromoMovie = (state) => {
+  return state[NAME_SPACE].promo;
+};
