+++ src/reducer/name-space.js
@@ -1,0 +1,7 @@
+const NameSpace = {
+  APP: `APP`,
+  DATA: `DATA`,
+  USER: `USER`,
+};
+
+export default NameSpace;
