+++ src/reducer/user/user.js
@@ -1,0 +1,32 @@
+import {extendObject} from "../../const";
+
+const AuthorizationStatus = {
+  AUTH: `AUTH`,
+  NO_AUTH: `NO_AUTH`,
+};
+
+const initialState = {
+  authorizationStatus: AuthorizationStatus.NO_AUTH,
+};
+
+const ActionType = {
+  REQUIRED_AUTHORIZATION: `REQUIRED_AUTHORIZATION`,
+};
+
+const ActionCreator = {
+  requireAuthorization: (status) => ({
+    type: ActionType.REQUIRED_AUTHORIZATION,
+    payload: status,
+  }),
+};
+
+const reducer = (state = initialState, action) => {
+  switch (action.type) {
+    case ActionType.REQUIRED_AUTHORIZATION:
+      return extendObject(state, {authorizationStatus: action.payload});
+  }
+
+  return state;
+};
+
+export {reducer, ActionType, ActionCreator, AuthorizationStatus};
